Обработка исключений и валидация, мелкие правки

+++ src/main/java/ru/kolyan/pathfinder/service/impl/SkillFeatServiceImpl.java
@@ -1,11 +1,13 @@
 package ru.kolyan.pathfinder.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import ru.kolyan.pathfinder.controller.skillfeat.request.CreateSkillFeatRequest;
 import ru.kolyan.pathfinder.controller.skillfeat.request.UpdateByIdSkillFeatRequest;
 import ru.kolyan.pathfinder.controller.skillfeat.response.GetAllSkillFeatResponse;
 import ru.kolyan.pathfinder.controller.skillfeat.response.GetByIdSkillFeatResponse;
+import ru.kolyan.pathfinder.exception.ConflictException;
 import ru.kolyan.pathfinder.exception.NotFoundException;
 import ru.kolyan.pathfinder.model.SkillFeat;
 import ru.kolyan.pathfinder.repository.SkillFeatRepository;
@@ -26,13 +28,19 @@ public class SkillFeatServiceImpl implements SkillFeatService {
         skillFeat.setName(request.getName());
         skillFeat.setDescription(request.getDescription());
 
-        skillFeatRepository.save(skillFeat);
+        try {
+            skillFeatRepository.save(skillFeat);
+        } catch (DataIntegrityViolationException e){
+            throw new ConflictException("Скил Фит с таким именем уже ест");
+        }
+
+
     }
 
     @Override
     public GetByIdSkillFeatResponse getById(UUID id) {
         SkillFeat skillFeat = skillFeatRepository.findById(id)
-                .orElseThrow(NotFoundException::new);
+                .orElseThrow(() -> new NotFoundException("Скил Фита с таким ID нету"));
 
         return GetByIdSkillFeatResponse.builder()
                 .id(skillFeat.getId())
@@ -58,14 +66,17 @@ public class SkillFeatServiceImpl implements SkillFeatService {
     }
 
     @Override
-    public void delete(UUID id) {
+    public void deleteById(UUID id) {
+        if(!skillFeatRepository.existsById(id)){
+            throw new NotFoundException("Скил Фита с таким ID нету");
+        }
         skillFeatRepository.deleteById(id);
     }
 
     @Override
     public void update(UpdateByIdSkillFeatRequest request, UUID id) {
         SkillFeat skillFeat = skillFeatRepository.findById(id)
-                .orElseThrow(NotFoundException::new);
+                .orElseThrow(() -> new NotFoundException("Скил Фита с таким ID нету"));
 
         Optional.ofNullable(request.getName())
                 .ifPresent(skillFeat::setName);
